Move client classes to udmi/lib package
