+++ pubber/src/main/java/daq/pubber/GatewayManager.java
@@ -6,14 +6,13 @@ import static com.google.udmi.util.GeneralUtils.ifNotNullGet;
 import static com.google.udmi.util.GeneralUtils.ifNotNullThen;
 import static com.google.udmi.util.GeneralUtils.ifNullThen;
 import static com.google.udmi.util.GeneralUtils.ifTrueThen;
-import static java.lang.String.format;
 import static java.util.Optional.ofNullable;
 import static udmi.schema.Category.GATEWAY_PROXY_TARGET;
 
 import com.google.udmi.util.SiteModel;
-import daq.pubber.client.GatewayManagerProvider;
-import daq.pubber.client.ProxyDeviceHostProvider;
 import java.util.Map;
+import udmi.lib.GatewayManagerProvider;
+import udmi.lib.ProxyDeviceHostProvider;
 import udmi.schema.Entry;
 import udmi.schema.GatewayConfig;
 import udmi.schema.GatewayState;
