+++ pubber/src/main/java/daq/pubber/LocalnetManager.java
@@ -1,9 +1,9 @@
 package daq.pubber;
 
-import daq.pubber.client.LocalnetManagerProvider;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Collectors;
+import udmi.lib.LocalnetManagerProvider;
 import udmi.schema.LocalnetConfig;
 import udmi.schema.LocalnetState;
 import udmi.schema.PubberConfiguration;
