+++ pubber/src/main/java/daq/pubber/MqttPublisher.java
@@ -22,7 +22,6 @@ import com.google.udmi.util.CertManager;
 import com.google.udmi.util.NanSerializer;
 import com.google.udmi.util.SiteModel;
 import com.google.udmi.util.SiteModel.ClientInfo;
-import daq.pubber.client.PubberHostProvider;
 import io.jsonwebtoken.JwtBuilder;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -57,6 +56,7 @@ import org.eclipse.paho.client.mqttv3.persist.MemoryPersistence;
 import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import udmi.lib.PubberHostProvider;
 import udmi.schema.Basic;
 import udmi.schema.Config;
 import udmi.schema.EndpointConfiguration.Transport;
