+++ pubber/src/main/java/daq/pubber/Pubber.java
@@ -31,13 +31,8 @@ import com.google.udmi.util.CertManager;
 import com.google.udmi.util.SchemaVersion;
 import com.google.udmi.util.SiteModel;
 import com.google.udmi.util.SiteModel.MetadataException;
-import daq.pubber.MqttPublisher.FakeTopic;
-import daq.pubber.MqttPublisher.InjectedMessage;
-import daq.pubber.MqttPublisher.InjectedState;
 import daq.pubber.MqttPublisher.PublisherException;
 import daq.pubber.PubSubClient.Bundle;
-import daq.pubber.client.DeviceManagerProvider;
-import daq.pubber.client.PubberHostProvider;
 import java.io.File;
 import java.io.PrintStream;
 import java.time.Duration;
@@ -55,6 +50,8 @@ import java.util.function.Function;
 import org.apache.http.ConnectionClosedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import udmi.lib.DeviceManagerProvider;
+import udmi.lib.PubberHostProvider;
 import udmi.schema.Config;
 import udmi.schema.DevicePersistent;
 import udmi.schema.EndpointConfiguration;
