+++ pubber/src/main/java/daq/pubber/SystemManager.java
@@ -6,15 +6,13 @@ import static com.google.udmi.util.GeneralUtils.isTrue;
 import static com.google.udmi.util.JsonUtil.isoConvert;
 import static java.lang.String.format;
 
-import com.google.common.collect.ImmutableMap;
-import daq.pubber.client.PubberHostProvider;
-import daq.pubber.client.SystemManagerProvider;
 import java.io.File;
 import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
+import udmi.lib.PubberHostProvider;
+import udmi.lib.SystemManagerProvider;
 import udmi.schema.Entry;
 import udmi.schema.Level;
 import udmi.schema.Operation.SystemMode;
