+++ pubber/src/main/java/udmi/lib/PubberHostProvider.java
@@ -1,4 +1,4 @@
-package daq.pubber.client;
+package udmi.lib;
 
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.udmi.util.GeneralUtils.catchToNull;
@@ -24,9 +24,7 @@ import static daq.pubber.MqttDevice.CONFIG_TOPIC;
 import static daq.pubber.MqttDevice.ERRORS_TOPIC;
 import static daq.pubber.MqttDevice.STATE_TOPIC;
 import static daq.pubber.MqttPublisher.DEFAULT_CONFIG_WAIT_SEC;
-import static daq.pubber.Pubber.SYSTEM_EVENT_TOPIC;
 import static java.lang.String.format;
-import static java.util.Objects.requireNonNull;
 import static java.util.Optional.ofNullable;
 import static udmi.schema.BlobsetConfig.SystemBlobsets.IOT_ENDPOINT_CONFIG;
 
@@ -42,8 +40,6 @@ import daq.pubber.MqttDevice;
 import daq.pubber.MqttPublisher.FakeTopic;
 import daq.pubber.MqttPublisher.InjectedMessage;
 import daq.pubber.MqttPublisher.InjectedState;
-import daq.pubber.client.PointsetManagerProvider.ExtraPointsetEvent;
-import daq.pubber.client.SystemManagerProvider.ExtraSystemState;
 import java.io.File;
 import java.lang.reflect.Field;
 import java.time.Duration;
@@ -61,6 +57,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.locks.Lock;
 import java.util.function.Consumer;
 import java.util.function.Function;
+import udmi.lib.PointsetManagerProvider.ExtraPointsetEvent;
+import udmi.lib.SystemManagerProvider.ExtraSystemState;
 import udmi.schema.BlobBlobsetConfig;
 import udmi.schema.BlobBlobsetConfig.BlobPhase;
 import udmi.schema.BlobBlobsetState;
@@ -70,16 +68,12 @@ import udmi.schema.Category;
 import udmi.schema.Config;
 import udmi.schema.DevicePersistent;
 import udmi.schema.DiscoveryEvents;
-import udmi.schema.DiscoveryState;
 import udmi.schema.EndpointConfiguration;
 import udmi.schema.Entry;
 import udmi.schema.Envelope.SubType;
-import udmi.schema.GatewayState;
 import udmi.schema.Level;
-import udmi.schema.LocalnetState;
 import udmi.schema.Operation.SystemMode;
 import udmi.schema.PointsetEvents;
-import udmi.schema.PointsetState;
 import udmi.schema.PubberConfiguration;
 import udmi.schema.PubberOptions;
 import udmi.schema.State;
